Adds aria information to accordion - trello ticket #454 (#12)
* Adds aria information to accordion - trello ticket #454

Each accordion previously was announced as 'open accordion'. They now announce the content that's the accordion title, and the screen reader user is informed when they expand and collapse. Still a couple of things left to do before it's finished:

- [ ] tidy up the ID
- [ ] figure out a solution for the content coming in from the admin panel (heading levels)

* Add attributes to rest of accordion code

There is another section of HTML for the accordions. This adds in all the necessary attributes to that too. It also fixes an issue where the wrong fields were being picked up for the IDs.

* Add in handling for up and down keys

The W3C accordion example suggests that for navigating accordions, you should allow the user to use the up and down and home and esc buttons. This adds in the JS which allows them to do this.

* Hyphenate attribute text