Adds screen reader text to pagination - trello ticket #463 (#10)
* Adds screen reader text to pagination - trello ticket #463

This adds text to the pagination links. I've added text to the previous and next links as well as the numbers.

* Add translation capability to pagination screen reader text

This allows the text available to screen reader users to be translated. I've added in the text domain to the style.css, which the text snippets reference. The structure of the overall sentences can't be changed, because of the way we need to pull in values from the PHP, but from a quick look at Welsh grammar in translations, this should be okay.

Whilst there is some repetition of words, I've kept the strings largely as they are without trying to shorten them. This should help translators understand the overall string, and lower the risk of translations being incorrect (for example if words need a declination, or needs to be in a certain position depending on the context).