Refactor About Us video to work as inline rather than modal
As the video was already being implimented via ACF, I've kept the ACF
approach but used ACF's oEmbed instead as that is the correct ACF way of
embeding video into a page. As the video was no longer a modal popup,
I've had to make a few CSS changes so that it works responsively inline.